# Conflicts:
#	app/src/main/java/com/example/android4/core/designsystem/theme/Type.kt

+++ app/src/main/java/com/example/android4/presentation/home/HomeNavigation.kt
@@ -16,7 +16,7 @@ fun NavController.navigateToHome(
 
 fun NavGraphBuilder.homeNavGraph(
     paddingValues: PaddingValues,
-    navigateToRecommendCourse: () -> Unit
+    navigateToRecommendCourse: (Int) -> Unit
 ) {
     composable<Home> {
         HomeScreen(
